requirements.txt
apache-airflow-providers-docker
apache-airflow[amazon]